Moved independent tests out of standalone (with some changes)
